// backend/src/apis/ZA/model/update-ticket-request.ts
export class UpdateTicketRequest {
    'ticketType': string;
    'citationNo': string;
    'country': string;
    'licensePlate': string;
    'courthouse': string;
    'violationDate': string;
    'violationCodes': string;
    'institutionId': string;
    'modelCode': string;
    'makeCode': string;
    'modelType': string;
    'vehicleMake': string;
    'vehicleYear': string;
    'ticketStatus': number;
    'city': string;
    'address': string;
    'violationId': string;
    'customerId': string;
    'ticketSystem': string;
    'transactionState': number;
    'amount': string;
    'currency': string;
    'id': string;
    'hasMembership': boolean;
    'finished': boolean;
    'questionsAndAnswers': string[][];
    'paymentRefNumber': string;
    'amountPaid': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'paymentRefNumber',
            baseName: 'paymentRefNumber',
            type: 'string',
        },
        {
            name: 'questionsAndAnswers',
            baseName: 'questionsAndAnswers',
            type: 'string[][]',
        },
        {
            name: 'ticketType',
            baseName: 'ticketType',
            type: 'string',
        },
        {
            name: 'citationNo',
            baseName: 'citationNo',
            type: 'string',
        },
        {
            name: 'country',
            baseName: 'country',
            type: 'string',
        },
        {
            name: 'licensePlate',
            baseName: 'licensePlate',
            type: 'string',
        },
        {
            name: 'courthouse',
            baseName: 'courthouse',
            type: 'string',
        },
        {
            name: 'violationDate',
            baseName: 'violationDate',
            type: 'string',
        },
        {
            name: 'violationCodes',
            baseName: 'violationCodes',
            type: 'string',
        },
        {
            name: 'institutionId',
            baseName: 'institutionId',
            type: 'string',
        },
        {
            name: 'modelCode',
            baseName: 'modelCode',
            type: 'string',
        },
        {
            name: 'makeCode',
            baseName: 'makeCode',
            type: 'string',
        },
        {
            name: 'modelType',
            baseName: 'modelType',
            type: 'string',
        },
        {
            name: 'vehicleMake',
            baseName: 'vehicleMake',
            type: 'string',
        },
        {
            name: 'vehicleYear',
            baseName: 'vehicleYear',
            type: 'string',
        },
        {
            name: 'ticketStatus',
            baseName: 'ticketStatus',
            type: 'number',
        },
        {
            name: 'city',
            baseName: 'city',
            type: 'string',
        },
        {
            name: 'address',
            baseName: 'address',
            type: 'string',
        },
        {
            name: 'violationId',
            baseName: 'violationId',
            type: 'string',
        },
        {
            name: 'customerId',
            baseName: 'customerId',
            type: 'string',
        },
        {
            name: 'ticketSystem',
            baseName: 'ticketSystem',
            type: 'string',
        },
        {
            name: 'transactionState',
            baseName: 'transactionState',
            type: 'number',
        },
        {
            name: 'amount',
            baseName: 'amount',
            type: 'string',
        },
        {
            name: 'currency',
            baseName: 'currency',
            type: 'string',
        },
        {
            name: 'id',
            baseName: 'id',
            type: 'string',
        },
        {
            name: 'hasMembership',
            baseName: 'hasMembership',
            type: 'boolean',
        },
        {
            name: 'finished',
            baseName: 'finished',
            type: 'boolean',
        },
        {
            name: 'amountPaid',
            baseName: 'amountPaid',
            type: 'number',
        },
    ];

    static getAttributeTypeMap() {
        return UpdateTicketRequest.attributeTypeMap;
    }
}